+++ Serious-Engine/Sources/Engine/Math/Object3D_CSG.cpp
@@ -325,13 +325,13 @@ void CObjectCSG::PolygonEdgesToBSPEdges(
     if (ope.ope_Backward) {
       // add bsp edge with reverse vertices
       abed[iEdge] = DOUBLEbspedge3D(*ope.ope_Edge->oed_Vertex1,
-        *ope.ope_Edge->oed_Vertex0, (ULONG)(size_t)ope.ope_Edge);
+        *ope.ope_Edge->oed_Vertex0, (size_t)ope.ope_Edge);
 
     // if it is not reversed
     } else{
       // add bsp edge with normal vertices
       abed[iEdge] = DOUBLEbspedge3D(*ope.ope_Edge->oed_Vertex0,
-        *ope.ope_Edge->oed_Vertex1, (ULONG)(size_t)ope.ope_Edge);
+        *ope.ope_Edge->oed_Vertex1, (size_t)ope.ope_Edge);
     }
   }
 
@@ -543,7 +543,7 @@ void CObjectCSG::DoCSGSplitting(
         oc_popoPortalB1B2 = NULL;
 
         // create a bsp polygon from first temporary array
-        DOUBLEbsppolygon3D bpoA(*itopoA->opo_Plane, abedRemaining, (ULONG)(size_t)itopoA->opo_Plane);
+        DOUBLEbsppolygon3D bpoA(*itopoA->opo_Plane, abedRemaining, (size_t)itopoA->opo_Plane);
 
         // create a BSP cutter for B's sector BSP and A's polygon
         DOUBLEbspcutter3D bcCutter(bpoA, *itoscB->osc_BSPTree.bt_pbnRoot);
